deleted schemas, moved everything from it to the models instead since they are basicly the same thing.
