--- app/schemas/user.py
@@ -1,14 +1,0 @@
-from beanie import PydanticObjectId
-from fastapi_users import schemas
-
-
-class UserRead(schemas.BaseUser[PydanticObjectId]):
-    pass
-
-
-class UserCreate(schemas.BaseUserCreate):
-    pass
-
-
-class UserUpdate(schemas.BaseUserUpdate):
-    pass
